refactor: move verifier logic from operation.go
closes #431
