+++ pkg/restapi/vc/controller_test.go
@@ -24,7 +24,7 @@ import (
 	"github.com/trustbloc/edge-service/pkg/restapi/vc/operation"
 )
 
-func TestIssuerController_New(t *testing.T) {
+func TestController_New(t *testing.T) {
 	t.Run("test success", func(t *testing.T) {
 		client := edv.NewMockEDVClient("test", nil, nil, []string{"testID"})
 
@@ -34,7 +34,7 @@ func TestIssuerController_New(t *testing.T) {
 		controller, err := New(&operation.Config{StoreProvider: memstore.NewProvider(),
 			Crypto:             &cryptomock.Crypto{},
 			KMSSecretsProvider: mem.NewProvider(), EDVClient: client, KeyManager: &kms.KeyManager{CreateKeyValue: kh},
-			VDRI: &vdrimock.MockVDRIRegistry{}, HostURL: "", Mode: "issuer"})
+			VDRI: &vdrimock.MockVDRIRegistry{}, HostURL: ""})
 		require.NoError(t, err)
 		require.NotNil(t, controller)
 	})
@@ -43,50 +43,14 @@ func TestIssuerController_New(t *testing.T) {
 		client := edv.NewMockEDVClient("test", nil, nil, []string{"testID"})
 		controller, err := New(&operation.Config{StoreProvider: &mockstore.Provider{
 			ErrOpenStoreHandle: fmt.Errorf("error open store")}, EDVClient: client,
-			VDRI: &vdrimock.MockVDRIRegistry{}, HostURL: "", Mode: "issuer"})
+			VDRI: &vdrimock.MockVDRIRegistry{}, HostURL: ""})
 		require.Error(t, err)
 		require.Contains(t, err.Error(), "error open store")
 		require.Nil(t, controller)
 	})
 }
 
-func TestVerifierController_New(t *testing.T) {
-	t.Run("test success", func(t *testing.T) {
-		client := edv.NewMockEDVClient("test", nil, nil, []string{"testID"})
-
-		kh, err := keyset.NewHandle(ecdhes.ECDHES256KWAES256GCMKeyTemplate())
-		require.NoError(t, err)
-
-		controller, err := New(&operation.Config{StoreProvider: memstore.NewProvider(),
-			Crypto:             &cryptomock.Crypto{},
-			KMSSecretsProvider: mem.NewProvider(), EDVClient: client, KeyManager: &kms.KeyManager{CreateKeyValue: kh},
-			VDRI: &vdrimock.MockVDRIRegistry{}, HostURL: "", Mode: "verifier"})
-		require.NoError(t, err)
-		require.NotNil(t, controller)
-	})
-
-	t.Run("test error", func(t *testing.T) {
-		client := edv.NewMockEDVClient("test", nil, nil, []string{"testID"})
-		controller, err := New(&operation.Config{StoreProvider: &mockstore.Provider{
-			ErrOpenStoreHandle: fmt.Errorf("error open store")}, EDVClient: client,
-			VDRI: &vdrimock.MockVDRIRegistry{}, HostURL: "", Mode: "verifier"})
-		require.Error(t, err)
-		require.Contains(t, err.Error(), "error open store")
-		require.Nil(t, controller)
-	})
-}
-
-func TestControllerInvalidMode_New(t *testing.T) {
-	t.Run("must return error if an invalid mode is given", func(t *testing.T) {
-		_, err := New(&operation.Config{StoreProvider: &mockstore.Provider{
-			ErrOpenStoreHandle: fmt.Errorf("error open store")},
-			EDVClient: edv.NewMockEDVClient("test", nil, nil, []string{"testID"}),
-			VDRI:      &vdrimock.MockVDRIRegistry{}, HostURL: "", Mode: "invalid"})
-		require.Error(t, err)
-	})
-}
-
-func TestIssuerController_GetOperations(t *testing.T) {
+func TestController_GetOperations(t *testing.T) {
 	client := edv.NewMockEDVClient("test", nil, nil, []string{"testID"})
 
 	kh, err := keyset.NewHandle(ecdhes.ECDHES256KWAES256GCMKeyTemplate())
@@ -95,7 +59,7 @@ func TestIssuerController_GetOperations(t *testing.T) {
 	controller, err := New(&operation.Config{StoreProvider: memstore.NewProvider(),
 		Crypto:             &cryptomock.Crypto{},
 		KMSSecretsProvider: mem.NewProvider(), EDVClient: client, KeyManager: &kms.KeyManager{CreateKeyValue: kh},
-		VDRI: &vdrimock.MockVDRIRegistry{}, HostURL: "", Mode: "issuer"})
+		VDRI: &vdrimock.MockVDRIRegistry{}, HostURL: ""})
 
 	require.NoError(t, err)
 	require.NotNil(t, controller)
@@ -103,23 +67,4 @@ func TestIssuerController_GetOperations(t *testing.T) {
 	ops := controller.GetOperations()
 
 	require.Equal(t, 9, len(ops))
-}
-
-func TestVerifierController_GetOperations(t *testing.T) {
-	client := edv.NewMockEDVClient("test", nil, nil, []string{"testID"})
-
-	kh, err := keyset.NewHandle(ecdhes.ECDHES256KWAES256GCMKeyTemplate())
-	require.NoError(t, err)
-
-	controller, err := New(&operation.Config{StoreProvider: memstore.NewProvider(),
-		Crypto:             &cryptomock.Crypto{},
-		KMSSecretsProvider: mem.NewProvider(), EDVClient: client, KeyManager: &kms.KeyManager{CreateKeyValue: kh},
-		VDRI: &vdrimock.MockVDRIRegistry{}, HostURL: "", Mode: "verifier"})
-
-	require.NoError(t, err)
-	require.NotNil(t, controller)
-
-	ops := controller.GetOperations()
-
-	require.Equal(t, 2, len(ops))
 }
