+++ pkg/restapi/vc/operation/openapi.go
@@ -81,54 +81,6 @@ type verifiableCredentialRes struct { // nolint: unused,deadcode
 	// in: body
 }
 
-// verifyCredentialReq model
-//
-// swagger:parameters verifyCredentialReq
-type verifyCredentialReq struct { // nolint: unused,deadcode
-	// in: body
-	Params CredentialsVerificationRequest
-}
-
-// verifyCredentialSuccessResp model
-//
-// swagger:response verifyCredentialSuccessResp
-type verifyCredentialSuccessResp struct { // nolint: unused,deadcode
-	// in: body
-	CredentialsVerificationSuccessResponse
-}
-
-// verifyCredentialFailureResp model
-//
-// swagger:response verifyCredentialFailureResp
-type verifyCredentialFailureResp struct { // nolint: unused,deadcode
-	// in: body
-	Checks []*CredentialsVerificationCheckResult `json:"checks,omitempty"`
-}
-
-// verifyCredentialReq model
-//
-// swagger:parameters verifyPresentationReq
-type verifyPresentationReq struct { // nolint: unused,deadcode
-	// in: body
-	Params VerifyPresentationRequest
-}
-
-// verifyPresentationSuccessResp model
-//
-// swagger:response verifyPresentationSuccessResp
-type verifyPresentationSuccessResp struct { // nolint: unused,deadcode
-	// in: body
-	VerifyPresentationSuccessResponse
-}
-
-// verifyPresentationFailureResp model
-//
-// swagger:response verifyPresentationFailureResp
-type verifyPresentationFailureResp struct { // nolint: unused,deadcode
-	// in: body
-	Checks []*VerifyPresentationCheckResult `json:"checks,omitempty"`
-}
-
 // generateKeypairResp model
 //
 // swagger:response generateKeypairResp
