+++ pkg/restapi/vc/operation/operations.go
@@ -14,16 +14,10 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"net/http"
 	"net/url"
 	"strings"
 	"time"
-
-	commondid "github.com/trustbloc/edge-service/pkg/restapi/internal/common/did"
-
-	commhttp "github.com/trustbloc/edge-service/pkg/restapi/internal/common/http"
-	"github.com/trustbloc/edge-service/pkg/restapi/model"
 
 	"github.com/btcsuite/btcutil/base58"
 	"github.com/google/tink/go/keyset"
@@ -45,6 +39,9 @@ import (
 	vcprofile "github.com/trustbloc/edge-service/pkg/doc/vc/profile"
 	cslstatus "github.com/trustbloc/edge-service/pkg/doc/vc/status/csl"
 	"github.com/trustbloc/edge-service/pkg/internal/common/support"
+	commondid "github.com/trustbloc/edge-service/pkg/restapi/internal/common/did"
+	commhttp "github.com/trustbloc/edge-service/pkg/restapi/internal/common/http"
+	"github.com/trustbloc/edge-service/pkg/restapi/model"
 )
 
 const (
@@ -64,37 +61,15 @@ const (
 	kmsBasePath                    = "/kms"
 	generateKeypairPath            = kmsBasePath + "/generatekeypair"
 
-	// verifier endpoints
-	verifierBasePath                  = "/verifier"
-	credentialsVerificationEndpoint   = verifierBasePath + "/credentials"
-	presentationsVerificationEndpoint = verifierBasePath + "/presentations"
-
-	successMsg = "success"
-	cslSize    = 50
+	cslSize = 50
 
 	invalidRequestErrMsg = "Invalid request"
-
-	// credential verification checks
-	proofCheck  = "proof"
-	statusCheck = "status"
 
 	// supported proof purpose
 	assertionMethod      = "assertionMethod"
 	authentication       = "authentication"
 	capabilityDelegation = "capabilityDelegation"
 	capabilityInvocation = "capabilityInvocation"
-
-	// modes
-	issuerMode   = "issuer"
-	verifierMode = "verifier"
-	holderMode   = "holder"
-	combinedMode = "combined"
-
-	// proof data keys
-	challenge          = "challenge"
-	domain             = "domain"
-	proofPurpose       = "proofPurpose"
-	verificationMethod = "verificationMethod"
 
 	jsonWebSignature2020Context = "https://trustbloc.github.io/context/vc/credentials-v1.jsonld"
 )
@@ -130,10 +105,6 @@ type keyManager interface {
 	kms.KeyManager
 	ExportPubKeyBytes(id string) ([]byte, error)
 	ImportPrivateKey(privKey interface{}, kt kms.KeyType, opts ...localkms.PrivateKeyOpts) (string, *keyset.Handle, error)
-}
-
-type httpClient interface {
-	Do(req *http.Request) (*http.Response, error)
 }
 
 type commonDID interface {
@@ -177,7 +148,6 @@ func New(config *Config) (*Operation, error) {
 		jweDecrypter:         jweDecrypter,
 		vcStatusManager:      vcStatusManager,
 		domain:               config.Domain,
-		httpClient:           &http.Client{Transport: &http.Transport{TLSClientConfig: config.TLSConfig}},
 		HostURL:              config.HostURL,
 		macKeyHandle:         kh,
 		macCrypto:            config.Crypto,
@@ -198,7 +168,6 @@ type Config struct {
 	VDRI               vdriapi.Registry
 	HostURL            string
 	Domain             string
-	Mode               string
 	TLSConfig          *tls.Config
 	Crypto             ariescrypto.Crypto
 }
@@ -214,7 +183,6 @@ type Operation struct {
 	jweDecrypter         jose.Decrypter
 	vcStatusManager      vcStatusManager
 	domain               string
-	httpClient           httpClient
 	HostURL              string
 	macKeyHandle         *keyset.Handle
 	macCrypto            ariescrypto.Crypto
@@ -223,34 +191,7 @@ type Operation struct {
 }
 
 // GetRESTHandlers get all controller API handler available for this service
-func (o *Operation) GetRESTHandlers(mode string) ([]Handler, error) {
-	switch mode {
-	case verifierMode:
-		return o.verifierHandlers(), nil
-	case issuerMode:
-		return o.issuerHandlers(), nil
-	case holderMode:
-		// TODO remove after separate issue and verifier
-		return nil, nil
-	case combinedMode:
-		vh := o.verifierHandlers()
-		ih := o.issuerHandlers()
-
-		return append(vh, ih...), nil
-	default:
-		return nil, fmt.Errorf("invalid operation mode: %s", mode)
-	}
-}
-
-func (o *Operation) verifierHandlers() []Handler {
-	return []Handler{
-		support.NewHTTPHandler(credentialsVerificationEndpoint, http.MethodPost, o.verifyCredentialHandler),
-		support.NewHTTPHandler(presentationsVerificationEndpoint, http.MethodPost,
-			o.verifyPresentationHandler),
-	}
-}
-
-func (o *Operation) issuerHandlers() []Handler {
+func (o *Operation) GetRESTHandlers() ([]Handler, error) {
 	return []Handler{
 		// issuer profile
 		support.NewHTTPHandler(createProfileEndpoint, http.MethodPost, o.createIssuerProfileHandler),
@@ -268,7 +209,7 @@ func (o *Operation) issuerHandlers() []Handler {
 		support.NewHTTPHandler(generateKeypairPath, http.MethodGet, o.generateKeypairHandler),
 		support.NewHTTPHandler(issueCredentialPath, http.MethodPost, o.issueCredentialHandler),
 		support.NewHTTPHandler(composeAndIssueCredentialPath, http.MethodPost, o.composeAndIssueCredentialHandler),
-	}
+	}, nil
 }
 
 // RetrieveCredentialStatus swagger:route GET /status/{id} issuer retrieveCredentialStatusReq
@@ -289,76 +230,6 @@ func (o *Operation) retrieveCredentialStatus(rw http.ResponseWriter, req *http.R
 
 	rw.WriteHeader(http.StatusOK)
 	commhttp.WriteResponse(rw, csl)
-}
-
-func (o *Operation) checkVCStatus(vclID, vcID string) (*VerifyCredentialResponse, error) {
-	vcResp := &VerifyCredentialResponse{
-		Verified: false}
-
-	req, err := http.NewRequest(http.MethodGet, vclID, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	resp, err := o.sendHTTPRequest(req, http.StatusOK)
-	if err != nil {
-		return nil, err
-	}
-
-	var csl cslstatus.CSL
-	if err := json.Unmarshal(resp, &csl); err != nil {
-		return nil, fmt.Errorf("failed to unmarshal resp to csl: %w", err)
-	}
-
-	for _, vcStatus := range csl.VC {
-		if !strings.Contains(vcStatus, vcID) {
-			continue
-		}
-
-		statusVc, err := o.parseAndVerifyVC([]byte(vcStatus))
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse and verify status vc: %s", err.Error())
-		}
-
-		subjectBytes, err := json.Marshal(statusVc.Subject)
-		if err != nil {
-			return nil, fmt.Errorf(fmt.Sprintf("failed to marshal status vc subject: %s", err.Error()))
-		}
-
-		vcResp.Message = string(subjectBytes)
-
-		return vcResp, nil
-	}
-
-	vcResp.Verified = true
-	vcResp.Message = successMsg
-
-	return vcResp, nil
-}
-
-func (o *Operation) sendHTTPRequest(req *http.Request, status int) ([]byte, error) {
-	resp, err := o.httpClient.Do(req)
-	if err != nil {
-		return nil, err
-	}
-
-	defer func() {
-		err = resp.Body.Close()
-		if err != nil {
-			log.Warn("failed to close response body")
-		}
-	}()
-
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		log.Warnf("failed to read response body for status %d: %s", resp.StatusCode, err)
-	}
-
-	if resp.StatusCode != status {
-		return nil, fmt.Errorf("failed to read response body for status %d: %s", resp.StatusCode, string(body))
-	}
-
-	return body, nil
 }
 
 // UpdateCredentialStatus swagger:route POST /updateStatus issuer updateCredentialStatusReq
@@ -1032,227 +903,6 @@ func (o *Operation) createKey(keyType kms.KeyType) (string, []byte, error) {
 	return keyID, pubKeyBytes, nil
 }
 
-// nolint dupl
-// VerifyCredential swagger:route POST /verifier/credentials verifier verifyCredentialReq
-//
-// Verifies a credential.
-//
-// Responses:
-//    default: genericError
-//        200: verifyCredentialSuccessResp
-//        400: verifyCredentialFailureResp
-func (o *Operation) verifyCredentialHandler(rw http.ResponseWriter, req *http.Request) {
-	// get the request
-	verificationReq := CredentialsVerificationRequest{}
-
-	err := json.NewDecoder(req.Body).Decode(&verificationReq)
-	if err != nil {
-		commhttp.WriteErrorResponse(rw, http.StatusBadRequest, fmt.Sprintf(invalidRequestErrMsg+": %s", err.Error()))
-
-		return
-	}
-
-	vc, err := verifiable.NewUnverifiedCredential(verificationReq.Credential)
-	if err != nil {
-		commhttp.WriteErrorResponse(rw, http.StatusBadRequest, fmt.Sprintf(invalidRequestErrMsg+": %s", err.Error()))
-
-		return
-	}
-
-	checks := []string{proofCheck}
-
-	// if req contains checks, then override the default checks
-	if verificationReq.Opts != nil && len(verificationReq.Opts.Checks) != 0 {
-		checks = verificationReq.Opts.Checks
-	}
-
-	var result []CredentialsVerificationCheckResult
-
-	for _, val := range checks {
-		switch val {
-		case proofCheck:
-			err := o.validateCredentialProof(verificationReq.Credential, verificationReq.Opts, false)
-			if err != nil {
-				result = append(result, CredentialsVerificationCheckResult{
-					Check: val,
-					Error: err.Error(),
-				})
-			}
-		case statusCheck:
-			failureMessage := ""
-			if vc.Status != nil && vc.Status.ID != "" {
-				ver, err := o.checkVCStatus(vc.Status.ID, vc.ID)
-
-				if err != nil {
-					failureMessage = fmt.Sprintf("failed to fetch the status : %s", err.Error())
-				} else if !ver.Verified {
-					failureMessage = ver.Message
-				}
-			}
-
-			if failureMessage != "" {
-				result = append(result, CredentialsVerificationCheckResult{
-					Check: val,
-					Error: failureMessage,
-				})
-			}
-		default:
-			result = append(result, CredentialsVerificationCheckResult{
-				Check: val,
-				Error: "check not supported",
-			})
-		}
-	}
-
-	if len(result) == 0 {
-		rw.WriteHeader(http.StatusOK)
-		commhttp.WriteResponse(rw, &CredentialsVerificationSuccessResponse{
-			Checks: checks,
-		})
-	} else {
-		rw.WriteHeader(http.StatusBadRequest)
-		commhttp.WriteResponse(rw, &CredentialsVerificationFailResponse{
-			Checks: result,
-		})
-	}
-}
-
-// VerifyPresentation swagger:route POST /verifier/presentations verifier verifyPresentationReq
-//
-// Verifies a presentation.
-//
-// Responses:
-//    default: genericError
-//        200: verifyPresentationSuccessResp
-//        400: verifyPresentationFailureResp
-func (o *Operation) verifyPresentationHandler(rw http.ResponseWriter, req *http.Request) {
-	// get the request
-	verificationReq := VerifyPresentationRequest{}
-
-	err := json.NewDecoder(req.Body).Decode(&verificationReq)
-	if err != nil {
-		commhttp.WriteErrorResponse(rw, http.StatusBadRequest, fmt.Sprintf(invalidRequestErrMsg+": %s", err.Error()))
-
-		return
-	}
-
-	checks := []string{proofCheck}
-
-	// if req contains checks, then override the default checks
-	if verificationReq.Opts != nil && len(verificationReq.Opts.Checks) != 0 {
-		checks = verificationReq.Opts.Checks
-	}
-
-	var result []VerifyPresentationCheckResult
-
-	for _, val := range checks {
-		switch val {
-		case proofCheck:
-			err := o.validatePresentationProof(verificationReq.Presentation, verificationReq.Opts)
-			if err != nil {
-				result = append(result, VerifyPresentationCheckResult{
-					Check: val,
-					Error: err.Error(),
-				})
-			}
-		default:
-			result = append(result, VerifyPresentationCheckResult{
-				Check: val,
-				Error: "check not supported",
-			})
-		}
-	}
-
-	if len(result) == 0 {
-		rw.WriteHeader(http.StatusOK)
-		commhttp.WriteResponse(rw, &VerifyPresentationSuccessResponse{
-			Checks: checks,
-		})
-	} else {
-		rw.WriteHeader(http.StatusBadRequest)
-		commhttp.WriteResponse(rw, &VerifyPresentationFailureResponse{
-			Checks: result,
-		})
-	}
-}
-
-func (o *Operation) validateCredentialProof(vcByte []byte, opts *CredentialsVerificationOptions,
-	vcInVPValidation bool) error {
-	vc, err := o.parseAndVerifyVCStrictMode(vcByte)
-
-	if err != nil {
-		return fmt.Errorf("verifiable credential proof validation error : %w", err)
-	}
-
-	if len(vc.Proofs) == 0 {
-		return errors.New("verifiable credential doesn't contains proof")
-	}
-
-	// validate proof challenge and domain
-	if opts == nil {
-		opts = &CredentialsVerificationOptions{}
-	}
-
-	// TODO https://github.com/trustbloc/edge-service/issues/412 figure out the process when vc has more than one proof
-	proof := vc.Proofs[0]
-
-	if !vcInVPValidation {
-		// validate challenge
-		if err := validateProofData(proof, challenge, opts.Challenge); err != nil {
-			return err
-		}
-
-		// validate domain
-		if err := validateProofData(proof, domain, opts.Domain); err != nil {
-			return err
-		}
-	}
-
-	// validate proof purpose
-	if err := validateProofPurpose(proof, o.vdri); err != nil {
-		return fmt.Errorf("verifiable credential proof purpose validation error : %w", err)
-	}
-
-	return nil
-}
-
-func (o *Operation) validatePresentationProof(vpByte []byte, opts *VerifyPresentationOptions) error {
-	vp, err := o.parseAndVerifyVP(vpByte)
-
-	if err != nil {
-		return fmt.Errorf("verifiable presentation proof validation error : %w", err)
-	}
-
-	// validate proof challenge and domain
-	if opts == nil {
-		opts = &VerifyPresentationOptions{}
-	}
-
-	var proof verifiable.Proof
-
-	// TODO https://github.com/trustbloc/edge-service/issues/412 figure out the process when vc has more than one proof
-	if len(vp.Proofs) != 0 {
-		proof = vp.Proofs[0]
-	}
-
-	// validate challenge
-	if err := validateProofData(proof, challenge, opts.Challenge); err != nil {
-		return err
-	}
-
-	// validate domain
-	if err := validateProofData(proof, domain, opts.Domain); err != nil {
-		return err
-	}
-
-	// validate proof purpose
-	if err := validateProofPurpose(proof, o.vdri); err != nil {
-		return fmt.Errorf("verifiable presentation proof purpose validation error : %w", err)
-	}
-
-	return nil
-}
-
 func (o *Operation) parseAndVerifyVC(vcBytes []byte) (*verifiable.Credential, error) {
 	vc, _, err := verifiable.NewCredential(
 		vcBytes,
@@ -1266,51 +916,6 @@ func (o *Operation) parseAndVerifyVC(vcBytes []byte) (*verifiable.Credential, er
 	}
 
 	return vc, nil
-}
-
-func (o *Operation) parseAndVerifyVCStrictMode(vcBytes []byte) (*verifiable.Credential, error) {
-	vc, _, err := verifiable.NewCredential(
-		vcBytes,
-		verifiable.WithPublicKeyFetcher(
-			verifiable.NewDIDKeyResolver(o.vdri).PublicKeyFetcher(),
-		),
-		verifiable.WithStrictValidation(),
-	)
-
-	if err != nil {
-		return nil, err
-	}
-
-	return vc, nil
-}
-
-func (o *Operation) parseAndVerifyVP(vpBytes []byte) (*verifiable.Presentation, error) {
-	vp, err := verifiable.NewPresentation(
-		vpBytes,
-		verifiable.WithPresPublicKeyFetcher(
-			verifiable.NewDIDKeyResolver(o.vdri).PublicKeyFetcher(),
-		),
-	)
-
-	if err != nil {
-		return nil, err
-	}
-	// vp is verified
-
-	// verify if the credentials in vp are valid
-	for _, cred := range vp.Credentials() {
-		vcBytes, err := json.Marshal(cred)
-		if err != nil {
-			return nil, err
-		}
-		// verify if the credential in vp is valid
-		err = o.validateCredentialProof(vcBytes, nil, true)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return vp, nil
 }
 
 func (o *Operation) queryVault(vaultID, vcID string) ([]string, error) {
@@ -1471,43 +1076,4 @@ func getDocIDFromURL(docURL string) string {
 	docIDToRetrieve := splitBySlashes[len(splitBySlashes)-1]
 
 	return docIDToRetrieve
-}
-
-func validateProofData(proof verifiable.Proof, key, expectedValue string) error {
-	actualVal := ""
-
-	val, ok := proof[key]
-	if ok {
-		actualVal, _ = val.(string) // nolint
-	}
-
-	if expectedValue != actualVal {
-		return fmt.Errorf("invalid %s in the proof : expected=%s actual=%s", key, expectedValue, actualVal)
-	}
-
-	return nil
-}
-
-func validateProofPurpose(proof verifiable.Proof, vdri vdriapi.Registry) error {
-	purposeVal, ok := proof[proofPurpose]
-	if !ok {
-		return errors.New("proof doesn't have purpose")
-	}
-
-	purpose, ok := purposeVal.(string)
-	if !ok {
-		return errors.New("proof purpose is not a string")
-	}
-
-	verificationMethodVal, ok := proof[verificationMethod]
-	if !ok {
-		return errors.New("proof doesn't have verification method")
-	}
-
-	verificationMethod, ok := verificationMethodVal.(string)
-	if !ok {
-		return errors.New("proof verification method is not a string")
-	}
-
-	return crypto.ValidateProofPurpose(purpose, verificationMethod, vdri)
 }
