+++ pkg/restapi/vc/operation/operations_test.go
@@ -13,22 +13,17 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
-	"strings"
 	"testing"
 	"time"
 
 	"github.com/btcsuite/btcutil/base58"
 	"github.com/google/tink/go/keyset"
-	"github.com/google/uuid"
 	"github.com/gorilla/mux"
 	"github.com/hyperledger/aries-framework-go/pkg/crypto/tinkcrypto/primitive/composite/ecdhes"
 	"github.com/hyperledger/aries-framework-go/pkg/doc/did"
 	"github.com/hyperledger/aries-framework-go/pkg/doc/jose"
-	"github.com/hyperledger/aries-framework-go/pkg/doc/signature/suite"
-	"github.com/hyperledger/aries-framework-go/pkg/doc/signature/suite/ed25519signature2018"
 	"github.com/hyperledger/aries-framework-go/pkg/doc/verifiable"
 	"github.com/hyperledger/aries-framework-go/pkg/framework/aries/api/vdri"
 	cryptomock "github.com/hyperledger/aries-framework-go/pkg/mock/crypto"
@@ -50,6 +45,8 @@ import (
 )
 
 const (
+	challenge = "challenge"
+	domain    = "domain"
 
 	// json keys for compose issue credential options
 	keyID   = "kid"
@@ -124,41 +121,6 @@ const (
 	  }
 	}`
 
-	prCardVC = `{
-	  "@context": [
-		"https://www.w3.org/2018/credentials/v1",
-		"https://w3id.org/citizenship/v1"
-	  ],
-	  "id": "https://issuer.oidp.uscis.gov/credentials/83627465",
-	  "type": [
-		"VerifiableCredential",
-		"PermanentResidentCard"
-	  ],
-	  "name": "Permanent Resident Card",
-	  "description": "Permanent Resident Card",
-	  "issuer": "did:example:28394728934792387",
-	  "issuanceDate": "2019-12-03T12:19:52Z",
-	  "expirationDate": "2029-12-03T12:19:52Z",
-	  "credentialSubject": {
-		"id": "did:example:b34ca6cd37bbf23",
-		"type": [
-		  "PermanentResident",
-		  "Person"
-		],
-		"givenName": "JOHN",
-		"familyName": "SMITH",
-		"gender": "Male",
-		"image": "data:image/png;base64,iVBORw0KGgo...kJggg==",
-		"residentSince": "2015-01-01",
-		"lprCategory": "C09",
-		"lprNumber": "999-999-999",
-		"commuterClassification": "C1",
-		"birthCountry": "Bahamas",
-		"birthDate": "1958-07-17"
-	  }
-	}
-	`
-
 	validVCWithoutStatus = `{` +
 		validContext + `,
 	  "id": "http://example.edu/credentials/1872",
@@ -173,35 +135,6 @@ const (
 	  "issuanceDate": "2010-01-01T19:23:24Z"
 	}`
 
-	validVCWithProof = `{
-	   "@context":[
-		  "https://www.w3.org/2018/credentials/v1"
-	   ],
-	   "credentialSchema":[
-	   ],
-	   "credentialStatus":{
-		  "id":"https://example.gov/status/24",
-		  "type":"CredentialStatusList2017"
-	   },
-	   "credentialSubject":{
-		  "id":"did:example:ebfeb1f712ebc6f1c276e12ec21"
-	   },
-	   "id":"http://example.edu/credentials/1872",
-	   "issuanceDate":"2010-01-01T19:23:24Z",
-	   "issuer":{
-		  "id":"did:example:76e12ec712ebc6f1c221ebfeb1f",
-		  "name":"Example University"
-	   },
-	   "proof":{
-		  "created":"2020-03-23T17:20:15Z",
-		  "jws":"eyJhbGciOiJFZDI1NTE5U2lnbmF0dXJlMjAxOCIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..` +
-		`-M4JuuSvRurmWqX0S_x2eXg-ZaaDhkAUQ1GlV9DjD0WKZUKJefSkSgevYro64aFzQEa_gK7b1akJCB7XZtH1Aw",
-		  "type":"Ed25519Signature2018",
-		  "verificationMethod":"did:trustbloc:testnet.trustbloc.local:EiBpn9XWyJlGpny_` +
-		`ViTH75fi43ThiIlGUyc1rQEb3VgreQ==#key-1"
-	   },
-	   "type":"VerifiableCredential"
-	}`
 	// VC without issuer
 	invalidVC = `{` +
 		validContext + `,
@@ -211,24 +144,6 @@ const (
 		"id": "did:example:ebfeb1f712ebc6f1c276e12ec21"
 	  },
 	  "issuanceDate": "2010-01-01T19:23:24Z"
-	}`
-
-	validVCStatus = `{
-	  "@context": [
-		"https://www.w3.org/2018/credentials/v1"
-	  ],
-	  "credentialSchema": [],
-	  "credentialSubject": {
-		"currentStatus": "Revoked",
-		"statusReason": "Disciplinary action"
-	  },
-	  "id": "#ID",
-	  "issuanceDate": "2020-02-18T17:55:31.1381994Z",
-	  "issuer": {
-		"id": "did:example:76e12ec712ebc6f1c221ebfeb12",
-		"name": "Example University"
-	  },
-	  "type": "VerifiableCredential"
 	}`
 
 	testStructuredDocMessage1 = `"Hello World!"`
@@ -251,81 +166,6 @@ const (
 	   "content":{
 		  "message":"Howdy World!"
 	   }
-	}`
-
-	validVP = `{
-		"@context": [
-			"https://www.w3.org/2018/credentials/v1",
-			"https://www.w3.org/2018/credentials/examples/v1"
-		],
-		"id": "urn:uuid:3978344f-8596-4c3a-a978-8fcaba3903c5",
-		"type": "VerifiablePresentation",
-		"verifiableCredential": [{
-			"@context": [
-				"https://www.w3.org/2018/credentials/v1",
-				"https://www.w3.org/2018/credentials/examples/v1"
-			],
-			"id": "http://example.edu/credentials/1872",
-			"type": "VerifiableCredential",
-			"credentialSubject": {
-				"id": "did:example:ebfeb1f712ebc6f1c276e12ec21"
-			},
-			"issuer": {
-				"id": "did:example:76e12ec712ebc6f1c221ebfeb1f",
-				"name": "Example University"
-			},
-			"issuanceDate": "2010-01-01T19:23:24Z",
-			"credentialStatus": {
-				"id": "https://example.gov/status/24",
-				"type": "CredentialStatusList2017"
-			}
-		}],
-		"holder": "did:example:ebfeb1f712ebc6f1c276e12ec21",
-		"proof": {
-			"type": "Ed25519Signature2018",
-			"created": "2020-01-21T16:44:53+02:00",
-			"proofValue": "eyJhbGciOiJSUzI1NiIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..kTCYt5XsITJX1CxPCT8yAV-TVIw5WEuts01m` +
-		`q-pQy7UJiN5mgREEMGlv50aqzpqh4Qq_PbChOMqsLfRoPsnsgxD-WUcX16dUOqV0G_zS245-kronKb78cPktb3rk-BuQy72IFLN25DYuNzVB` +
-		`Ah4vGHSrQyHUGlcTwLtjPAnKb78"
-		},
-		"refreshService": {
-			"id": "https://example.edu/refresh/3732",
-			"type": "ManualRefreshService2018"
-		}
-	}`
-
-	vpWithoutProof = `{
-		"@context": [
-			"https://www.w3.org/2018/credentials/v1",
-			"https://www.w3.org/2018/credentials/examples/v1"
-		],
-		"id": "urn:uuid:3978344f-8596-4c3a-a978-8fcaba3903c5",
-		"type": "VerifiablePresentation",
-		"verifiableCredential": [{
-			"@context": [
-				"https://www.w3.org/2018/credentials/v1",
-				"https://www.w3.org/2018/credentials/examples/v1"
-			],
-			"id": "http://example.edu/credentials/1872",
-			"type": "VerifiableCredential",
-			"credentialSubject": {
-				"id": "did:example:ebfeb1f712ebc6f1c276e12ec21"
-			},
-			"issuer": {
-				"id": "did:example:76e12ec712ebc6f1c221ebfeb1f",
-				"name": "Example University"
-			},
-			"issuanceDate": "2010-01-01T19:23:24Z",
-			"credentialStatus": {
-				"id": "https://example.gov/status/24",
-				"type": "CredentialStatusList2017"
-			}
-		}],
-		"holder": "did:example:ebfeb1f712ebc6f1c276e12ec21",
-		"refreshService": {
-			"id": "https://example.edu/refresh/3732",
-			"type": "ManualRefreshService2018"
-		}
 	}`
 )
 
@@ -364,20 +204,14 @@ func TestNew(t *testing.T) {
 }
 
 func TestUpdateCredentialStatusHandler(t *testing.T) {
-	const (
-		issuerMode   = "issuer"
-		combinedMode = "combined"
-	)
-
-	testUpdateCredentialStatusHandler(t, issuerMode)
-	testUpdateCredentialStatusHandler(t, combinedMode)
+	testUpdateCredentialStatusHandler(t)
 }
 
-func testUpdateCredentialStatusHandler(t *testing.T, mode string) {
+func testUpdateCredentialStatusHandler(t *testing.T) {
 	client := edv.NewMockEDVClient("test", nil, nil, []string{"testID"})
 	s := make(map[string][]byte)
-	s["profile_"+issuerMode+"_Example University"] = []byte(testIssuerProfile)
-	s["profile_"+issuerMode+"_vc without status"] = []byte(testIssuerProfileWithDisableVCStatus)
+	s["profile_issuer_Example University"] = []byte(testIssuerProfile)
+	s["profile_issuer_vc without status"] = []byte(testIssuerProfileWithDisableVCStatus)
 
 	kh, err := keyset.NewHandle(ecdhes.ECDHES256KWAES256GCMKeyTemplate())
 	require.NoError(t, err)
@@ -390,7 +224,7 @@ func testUpdateCredentialStatusHandler(t *testing.T, mode string) {
 
 	op.vcStatusManager = &mockVCStatusManager{getCSLValue: &cslstatus.CSL{}}
 
-	updateCredentialStatusHandler := getHandler(t, op, updateCredentialStatusEndpoint, mode)
+	updateCredentialStatusHandler := getHandler(t, op, updateCredentialStatusEndpoint)
 
 	t.Run("update credential status success", func(t *testing.T) {
 		ucsReq := UpdateCredentialStatusRequest{Credential: validVC, Status: "revoked"}
@@ -458,7 +292,7 @@ func testUpdateCredentialStatusHandler(t *testing.T, mode string) {
 			Crypto:             &cryptomock.Crypto{}, VDRI: &vdrimock.MockVDRIRegistry{}, HostURL: "localhost:8080"})
 		require.NoError(t, err)
 		op.vcStatusManager = &mockVCStatusManager{getCSLValue: &cslstatus.CSL{}}
-		updateCredentialStatusHandler := getHandler(t, op, updateCredentialStatusEndpoint, mode)
+		updateCredentialStatusHandler := getHandler(t, op, updateCredentialStatusEndpoint)
 
 		ucsReq := UpdateCredentialStatusRequest{Credential: validVC, Status: "revoked"}
 		ucsReqBytes, err := json.Marshal(ucsReq)
@@ -476,7 +310,7 @@ func testUpdateCredentialStatusHandler(t *testing.T, mode string) {
 
 	t.Run("test error from update vc status", func(t *testing.T) {
 		s := make(map[string][]byte)
-		s["profile_"+issuerMode+"_Example University"] = []byte(testIssuerProfile)
+		s["profile_issuer_Example University"] = []byte(testIssuerProfile)
 
 		kh, err := keyset.NewHandle(ecdhes.ECDHES256KWAES256GCMKeyTemplate())
 		require.NoError(t, err)
@@ -488,7 +322,7 @@ func testUpdateCredentialStatusHandler(t *testing.T, mode string) {
 			Crypto:             &cryptomock.Crypto{}, VDRI: &vdrimock.MockVDRIRegistry{}, HostURL: "localhost:8080"})
 		require.NoError(t, err)
 		op.vcStatusManager = &mockVCStatusManager{updateVCStatusErr: fmt.Errorf("error update vc status")}
-		updateCredentialStatusHandler := getHandler(t, op, updateCredentialStatusEndpoint, mode)
+		updateCredentialStatusHandler := getHandler(t, op, updateCredentialStatusEndpoint)
 
 		ucsReq := UpdateCredentialStatusRequest{Credential: validVC, Status: "revoked"}
 		ucsReqBytes, err := json.Marshal(ucsReq)
@@ -506,13 +340,7 @@ func testUpdateCredentialStatusHandler(t *testing.T, mode string) {
 }
 
 func TestCreateProfileHandler(t *testing.T) {
-	const (
-		issuerMode   = "issuer"
-		combinedMode = "combined"
-	)
-
-	testCreateProfileHandler(t, issuerMode)
-	testCreateProfileHandler(t, combinedMode)
+	testCreateProfileHandler(t)
 }
 
 type mockCommonDID struct {
@@ -526,7 +354,7 @@ func (m *mockCommonDID) CreateDID(keyType, signatureType, didID, privateKey, key
 	return m.createDIDValue, m.createDIDKeyID, m.createDIDErr
 }
 
-func testCreateProfileHandler(t *testing.T, mode string) {
+func testCreateProfileHandler(t *testing.T) {
 	kh, err := keyset.NewHandle(ecdhes.ECDHES256KWAES256GCMKeyTemplate())
 	require.NoError(t, err)
 
@@ -542,7 +370,7 @@ func testCreateProfileHandler(t *testing.T, mode string) {
 
 	op.commonDID = &mockCommonDID{}
 
-	createProfileHandler := getHandler(t, op, createProfileEndpoint, mode)
+	createProfileHandler := getHandler(t, op, createProfileEndpoint)
 
 	var logContents bytes.Buffer
 
@@ -582,7 +410,7 @@ func testCreateProfileHandler(t *testing.T, mode string) {
 
 		require.NoError(t, err)
 
-		createProfileHandler = getHandler(t, op, createProfileEndpoint, mode)
+		createProfileHandler = getHandler(t, op, createProfileEndpoint)
 
 		req, err := http.NewRequest(http.MethodPost, createProfileEndpoint,
 			bytes.NewBuffer([]byte(testIssuerProfileWithDID)))
@@ -616,7 +444,7 @@ func testCreateProfileHandler(t *testing.T, mode string) {
 
 		require.NoError(t, err)
 
-		createProfileHandler = getHandler(t, op, createProfileEndpoint, mode)
+		createProfileHandler = getHandler(t, op, createProfileEndpoint)
 
 		req, err := http.NewRequest(http.MethodPost, createProfileEndpoint,
 			bytes.NewBuffer([]byte(testIssuerProfileWithDID)))
@@ -687,7 +515,7 @@ func TestGetProfileHandler(t *testing.T) {
 
 	op.commonDID = &mockCommonDID{}
 
-	getProfileHandler := getHandler(t, op, getProfileEndpoint, "issuer")
+	getProfileHandler := getHandler(t, op, getProfileEndpoint)
 
 	var logContents bytes.Buffer
 
@@ -744,7 +572,7 @@ func createProfileSuccess(t *testing.T, op *Operation) *vcprofile.DataProfile {
 
 	rr := httptest.NewRecorder()
 
-	createProfileEndpoint := getHandler(t, op, createProfileEndpoint, "issuer")
+	createProfileEndpoint := getHandler(t, op, createProfileEndpoint)
 	createProfileEndpoint.Handle().ServeHTTP(rr, req)
 
 	profile := &vcprofile.DataProfile{}
@@ -1237,7 +1065,7 @@ func TestRetrieveVCHandler(t *testing.T) {
 
 		setMockEDVClientReadDocumentReturnValue(t, client, op, testStructuredDocument1)
 
-		retrieveVCHandler := getHandler(t, op, retrieveCredentialEndpoint, "issuer")
+		retrieveVCHandler := getHandler(t, op, retrieveCredentialEndpoint)
 
 		req, err := http.NewRequest(http.MethodGet, retrieveCredentialEndpoint,
 			bytes.NewBuffer([]byte(nil)))
@@ -1324,8 +1152,6 @@ func TestRetrieveVCHandler(t *testing.T) {
 }
 
 func TestVCStatus(t *testing.T) {
-	const mode = "issuer"
-
 	t.Run("test error from get CSL", func(t *testing.T) {
 		client := edv.NewMockEDVClient("test", nil, nil, []string{"testID"})
 
@@ -1343,7 +1169,7 @@ func TestVCStatus(t *testing.T) {
 
 		op.vcStatusManager = &mockVCStatusManager{getCSLErr: fmt.Errorf("error get csl")}
 
-		vcStatusHandler := getHandler(t, op, credentialStatusEndpoint, mode)
+		vcStatusHandler := getHandler(t, op, credentialStatusEndpoint)
 
 		req, err := http.NewRequest(http.MethodGet, credentialStatus+"/1", nil)
 		require.NoError(t, err)
@@ -1372,7 +1198,7 @@ func TestVCStatus(t *testing.T) {
 		op.vcStatusManager = &mockVCStatusManager{
 			getCSLValue: &cslstatus.CSL{ID: "https://example.gov/status/24", VC: []string{}}}
 
-		vcStatusHandler := getHandler(t, op, credentialStatusEndpoint, mode)
+		vcStatusHandler := getHandler(t, op, credentialStatusEndpoint)
 
 		req, err := http.NewRequest(http.MethodGet, credentialStatus+"/1", nil)
 		require.NoError(t, err)
@@ -1430,27 +1256,17 @@ func TestOperation_GetRESTHandlers(t *testing.T) {
 	op, err := New(&Config{StoreProvider: memstore.NewProvider(),
 		KMSSecretsProvider: mem.NewProvider(),
 		Crypto:             &cryptomock.Crypto{},
-		EDVClient:          edv.NewMockEDVClient("test", nil, nil, []string{"testID"}),
-		KeyManager:         &kms.KeyManager{CreateKeyValue: kh},
-		VDRI:               &vdrimock.MockVDRIRegistry{},
-		HostURL:            "localhost:8080"})
+		EDVClient: edv.NewMockEDVClient("test",
+			nil, nil, []string{"testID"}),
+		KeyManager: &kms.KeyManager{CreateKeyValue: kh},
+		VDRI:       &vdrimock.MockVDRIRegistry{},
+		HostURL:    "localhost:8080"})
 
 	require.NoError(t, err)
 
-	t.Run("invalid mode", func(t *testing.T) {
-		_, err := op.GetRESTHandlers("invalid")
-		require.Error(t, err)
-	})
-	t.Run("issuer mode", func(t *testing.T) {
-		handlers, err := op.GetRESTHandlers("issuer")
-		require.NoError(t, err)
-		require.NotEmpty(t, handlers)
-	})
-	t.Run("verifier mode", func(t *testing.T) {
-		handlers, err := op.GetRESTHandlers("verifier")
-		require.NoError(t, err)
-		require.NotEmpty(t, handlers)
-	})
+	handlers, err := op.GetRESTHandlers()
+	require.NoError(t, err)
+	require.NotEmpty(t, handlers)
 }
 
 func TestIssueCredential(t *testing.T) {
@@ -1484,7 +1300,7 @@ func TestIssueCredential(t *testing.T) {
 	urlVars := make(map[string]string)
 	urlVars[profileIDPathParam] = profile.Name
 
-	handler := getHandler(t, op, issueCredentialPath, issuerMode)
+	handler := getHandler(t, op, issueCredentialPath)
 
 	t.Run("issue credential - success", func(t *testing.T) {
 		pubKey, _, err := ed25519.GenerateKey(rand.Reader)
@@ -1516,7 +1332,7 @@ func TestIssueCredential(t *testing.T) {
 		err = ops.profileStore.SaveProfile(profile)
 		require.NoError(t, err)
 
-		issueCredentialHandler := getHandler(t, ops, issueCredentialPath, issuerMode)
+		issueCredentialHandler := getHandler(t, ops, issueCredentialPath)
 
 		const createdTime = "2011-04-16T18:11:09-04:00"
 		ct, err := time.Parse(time.RFC3339, createdTime)
@@ -1636,7 +1452,7 @@ func TestIssueCredential(t *testing.T) {
 		err = ops.profileStore.SaveProfile(profile)
 		require.NoError(t, err)
 
-		issueCredentialHandler := getHandler(t, ops, issueCredentialPath, issuerMode)
+		issueCredentialHandler := getHandler(t, ops, issueCredentialPath)
 
 		req := &IssueCredentialRequest{
 			Credential: []byte(validVC),
@@ -1698,7 +1514,7 @@ func TestIssueCredential(t *testing.T) {
 		err = ops.profileStore.SaveProfile(profile)
 		require.NoError(t, err)
 
-		issueCredentialHandler := getHandler(t, ops, issueCredentialPath, issuerMode)
+		issueCredentialHandler := getHandler(t, ops, issueCredentialPath)
 
 		req := &IssueCredentialRequest{
 			Credential: []byte(validVC),
@@ -1728,7 +1544,7 @@ func TestIssueCredential(t *testing.T) {
 		})
 		require.NoError(t, err)
 
-		issueCredentialHandler := getHandler(t, ops, issueCredentialPath, issuerMode)
+		issueCredentialHandler := getHandler(t, ops, issueCredentialPath)
 
 		rr := serveHTTPMux(t, issueCredentialHandler, endpoint, nil, urlVars)
 
@@ -1826,7 +1642,7 @@ func TestIssueCredential(t *testing.T) {
 		})
 		require.NoError(t, err)
 
-		issueHandler := getHandler(t, op1, issueCredentialPath, issuerMode)
+		issueHandler := getHandler(t, op1, issueCredentialPath)
 
 		req := &IssueCredentialRequest{
 			Credential: []byte(validVC),
@@ -1863,7 +1679,7 @@ func TestIssueCredential(t *testing.T) {
 
 		op.vcStatusManager = &mockCredentialStatusManager{CreateErr: errors.New("csl error")}
 
-		issueCredentialHandler := getHandler(t, op, issueCredentialPath, issuerMode)
+		issueCredentialHandler := getHandler(t, op, issueCredentialPath)
 
 		req := &IssueCredentialRequest{
 			Credential: []byte(validVC),
@@ -1897,7 +1713,7 @@ func TestIssueCredential(t *testing.T) {
 		err = op.profileStore.SaveProfile(profile)
 		require.NoError(t, err)
 
-		issueCredentialHandler := getHandler(t, op, issueCredentialPath, issuerMode)
+		issueCredentialHandler := getHandler(t, op, issueCredentialPath)
 
 		req := &IssueCredentialRequest{
 			Credential: []byte(validVC),
@@ -1932,7 +1748,7 @@ func TestIssueCredential(t *testing.T) {
 		err = op.profileStore.SaveProfile(profile)
 		require.NoError(t, err)
 
-		issueCredentialHandler := getHandler(t, op, issueCredentialPath, issuerMode)
+		issueCredentialHandler := getHandler(t, op, issueCredentialPath)
 
 		req := &IssueCredentialRequest{
 			Credential: []byte(validVC),
@@ -1998,7 +1814,7 @@ func TestComposeAndIssueCredential(t *testing.T) {
 	})
 	require.NoError(t, err)
 
-	handler := getHandler(t, op, composeAndIssueCredentialPath, issuerMode)
+	handler := getHandler(t, op, composeAndIssueCredentialPath)
 
 	endpoint := "/test/credentials/composeAndIssueCredential"
 	issuerProfileDIDKey := "did:test:abc#" + key1ID
@@ -2033,7 +1849,7 @@ func TestComposeAndIssueCredential(t *testing.T) {
 		err = op.profileStore.SaveProfile(profile)
 		require.NoError(t, err)
 
-		restHandler := getHandler(t, op, composeAndIssueCredentialPath, issuerMode)
+		restHandler := getHandler(t, op, composeAndIssueCredentialPath)
 
 		claimJSON, err := json.Marshal(claim)
 		require.NoError(t, err)
@@ -2156,7 +1972,7 @@ func TestComposeAndIssueCredential(t *testing.T) {
 		})
 		require.NoError(t, err)
 
-		restHandler := getHandler(t, ops, composeAndIssueCredentialPath, issuerMode)
+		restHandler := getHandler(t, ops, composeAndIssueCredentialPath)
 
 		rr := serveHTTPMux(t, restHandler, endpoint, nil, urlVars)
 
@@ -2190,7 +2006,7 @@ func TestComposeAndIssueCredential(t *testing.T) {
 		reqBytes, err := json.Marshal(req)
 		require.NoError(t, err)
 
-		restHandler := getHandler(t, ops, composeAndIssueCredentialPath, issuerMode)
+		restHandler := getHandler(t, ops, composeAndIssueCredentialPath)
 
 		// invoke the endpoint
 		rr := serveHTTPMux(t, restHandler, endpoint, reqBytes, urlVars)
@@ -2278,7 +2094,7 @@ func TestComposeAndIssueCredential(t *testing.T) {
 		err = op1.profileStore.SaveProfile(profile)
 		require.NoError(t, err)
 
-		handler1 := getHandler(t, op1, composeAndIssueCredentialPath, issuerMode)
+		handler1 := getHandler(t, op1, composeAndIssueCredentialPath)
 
 		// invoke the endpoint
 		rr := serveHTTPMux(t, handler1, endpoint, reqBytes, urlVars)
@@ -2416,7 +2232,7 @@ func TestGenerateKeypair(t *testing.T) {
 		})
 		require.NoError(t, err)
 
-		generateKeypairHandler := getHandler(t, op, generateKeypairPath, issuerMode)
+		generateKeypairHandler := getHandler(t, op, generateKeypairPath)
 
 		rr := serveHTTP(t, generateKeypairHandler.Handle(), http.MethodGet, generateKeypairPath, nil)
 
@@ -2442,831 +2258,13 @@ func TestGenerateKeypair(t *testing.T) {
 		require.NoError(t, err)
 		op.kms = &kms.KeyManager{CreateKeyErr: errors.New("kms - create keyset error")}
 
-		generateKeypairHandler := getHandler(t, op, generateKeypairPath, issuerMode)
+		generateKeypairHandler := getHandler(t, op, generateKeypairPath)
 
 		rr := serveHTTP(t, generateKeypairHandler.Handle(), http.MethodGet, generateKeypairPath, nil)
 
 		require.Equal(t, http.StatusInternalServerError, rr.Code)
 		require.Contains(t, rr.Body.String(), "failed to create key pair")
 	})
-}
-
-func TestCredentialVerifications(t *testing.T) {
-	vc, err := verifiable.NewUnverifiedCredential([]byte(prCardVC))
-	require.NoError(t, err)
-
-	vc.Context = append(vc.Context, cslstatus.Context)
-
-	kh, err := keyset.NewHandle(ecdhes.ECDHES256KWAES256GCMKeyTemplate())
-	require.NoError(t, err)
-
-	op, err := New(&Config{
-		Crypto:             &cryptomock.Crypto{},
-		StoreProvider:      memstore.NewProvider(),
-		KMSSecretsProvider: mem.NewProvider(),
-		KeyManager:         &kms.KeyManager{CreateKeyValue: kh},
-		VDRI:               &vdrimock.MockVDRIRegistry{},
-	})
-	require.NoError(t, err)
-
-	endpoint := credentialsVerificationEndpoint
-	didID := "did:test:EiBNfNRaz1Ll8BjVsbNv-fWc7K_KIoPuW8GFCh1_Tz_Iuw=="
-
-	verificationsHandler := getHandler(t, op, endpoint, verifierMode)
-
-	t.Run("credential verification - success", func(t *testing.T) {
-		pubKey, privKey, err := ed25519.GenerateKey(rand.Reader)
-		require.NoError(t, err)
-
-		didDoc := createDIDDoc(didID, pubKey)
-		verificationMethod := didDoc.PublicKey[0].ID
-
-		ops, err := New(&Config{
-			Crypto:             &cryptomock.Crypto{},
-			StoreProvider:      memstore.NewProvider(),
-			KMSSecretsProvider: mem.NewProvider(),
-			KeyManager:         &kms.KeyManager{CreateKeyValue: kh},
-			VDRI:               &vdrimock.MockVDRIRegistry{ResolveValue: didDoc},
-		})
-		require.NoError(t, err)
-
-		cslBytes, err := json.Marshal(&cslstatus.CSL{})
-		require.NoError(t, err)
-
-		ops.httpClient = &mockHTTPClient{doValue: &http.Response{StatusCode: http.StatusOK,
-			Body: ioutil.NopCloser(strings.NewReader(string(cslBytes)))}}
-
-		vc.Status = &verifiable.TypedID{
-			ID:   "http://example.com/status/100",
-			Type: "CredentialStatusList2017",
-		}
-
-		vcBytes, err := vc.MarshalJSON()
-		require.NoError(t, err)
-
-		// verify credential
-		handler := getHandler(t, ops, endpoint, verifierMode)
-
-		vReq := &CredentialsVerificationRequest{
-			Credential: getSignedVC(t, privKey, string(vcBytes), verificationMethod, domain, challenge),
-			Opts: &CredentialsVerificationOptions{
-				Checks:    []string{proofCheck, statusCheck},
-				Challenge: challenge,
-				Domain:    domain,
-			},
-		}
-
-		vReqBytes, err := json.Marshal(vReq)
-		require.NoError(t, err)
-
-		rr := serveHTTP(t, handler.Handle(), http.MethodPost, endpoint, vReqBytes)
-
-		require.Equal(t, http.StatusOK, rr.Code)
-
-		verificationResp := &CredentialsVerificationSuccessResponse{}
-		err = json.Unmarshal(rr.Body.Bytes(), &verificationResp)
-		require.NoError(t, err)
-		require.Equal(t, 2, len(verificationResp.Checks))
-	})
-
-	t.Run("credential verification - request doesn't contain checks", func(t *testing.T) {
-		req := &CredentialsVerificationRequest{
-			Credential: []byte(prCardVC),
-		}
-
-		reqBytes, err := json.Marshal(req)
-		require.NoError(t, err)
-
-		rr := serveHTTP(t, verificationsHandler.Handle(), http.MethodPost, endpoint, reqBytes)
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-
-		// verify that the default check was performed
-		verificationResp := &CredentialsVerificationFailResponse{}
-		err = json.Unmarshal(rr.Body.Bytes(), &verificationResp)
-		require.NoError(t, err)
-		require.Equal(t, 1, len(verificationResp.Checks))
-		require.Equal(t, proofCheck, verificationResp.Checks[0].Check)
-		require.Equal(t, "verifiable credential doesn't contains proof", verificationResp.Checks[0].Error)
-	})
-
-	t.Run("credential verification - invalid credential", func(t *testing.T) {
-		req := &CredentialsVerificationRequest{
-			Credential: []byte(testIssuerProfileWithDID),
-		}
-
-		reqBytes, err := json.Marshal(req)
-		require.NoError(t, err)
-
-		rr := serveHTTP(t, verificationsHandler.Handle(), http.MethodPost, endpoint, reqBytes)
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-		require.Contains(t, rr.Body.String(), "Invalid request: build new credential")
-	})
-
-	t.Run("credential verification - proof check failure", func(t *testing.T) {
-		// no proof in VC
-		req := &CredentialsVerificationRequest{
-			Credential: []byte(prCardVC),
-			Opts: &CredentialsVerificationOptions{
-				Checks: []string{proofCheck},
-			},
-		}
-
-		reqBytes, err := json.Marshal(req)
-		require.NoError(t, err)
-
-		rr := serveHTTP(t, verificationsHandler.Handle(), http.MethodPost, endpoint, reqBytes)
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-
-		verificationResp := &CredentialsVerificationFailResponse{}
-		err = json.Unmarshal(rr.Body.Bytes(), &verificationResp)
-		require.NoError(t, err)
-		require.Equal(t, 1, len(verificationResp.Checks))
-		require.Equal(t, proofCheck, verificationResp.Checks[0].Check)
-		require.Equal(t, "verifiable credential doesn't contains proof", verificationResp.Checks[0].Error)
-
-		// proof validation error (DID not found)
-		req = &CredentialsVerificationRequest{
-			Credential: []byte(validVCWithProof),
-			Opts: &CredentialsVerificationOptions{
-				Checks: []string{proofCheck},
-			},
-		}
-
-		reqBytes, err = json.Marshal(req)
-		require.NoError(t, err)
-
-		rr = serveHTTP(t, verificationsHandler.Handle(), http.MethodPost, endpoint, reqBytes)
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-
-		verificationResp = &CredentialsVerificationFailResponse{}
-		err = json.Unmarshal(rr.Body.Bytes(), &verificationResp)
-		require.NoError(t, err)
-		require.Equal(t, 1, len(verificationResp.Checks))
-		require.Equal(t, proofCheck, verificationResp.Checks[0].Check)
-		require.Contains(t, verificationResp.Checks[0].Error, "verifiable credential proof validation error")
-	})
-
-	t.Run("credential verification - status check failure", func(t *testing.T) {
-		t.Run("status check failure - error fetching status", func(t *testing.T) {
-			vc.Status = &verifiable.TypedID{
-				ID: "http://example.com/status/100",
-			}
-
-			vcBytes, err := vc.MarshalJSON()
-			require.NoError(t, err)
-
-			req := &CredentialsVerificationRequest{
-				Credential: vcBytes,
-				Opts: &CredentialsVerificationOptions{
-					Checks: []string{statusCheck},
-				},
-			}
-
-			reqBytes, err := json.Marshal(req)
-			require.NoError(t, err)
-
-			rr := serveHTTP(t, verificationsHandler.Handle(), http.MethodPost, endpoint, reqBytes)
-
-			require.Equal(t, http.StatusBadRequest, rr.Code)
-
-			verificationResp := &CredentialsVerificationFailResponse{}
-			err = json.Unmarshal(rr.Body.Bytes(), &verificationResp)
-			require.NoError(t, err)
-			require.Equal(t, 1, len(verificationResp.Checks))
-			require.Equal(t, statusCheck, verificationResp.Checks[0].Check)
-			require.Contains(t, verificationResp.Checks[0].Error, "failed to fetch the status")
-		})
-
-		t.Run("status check failure - revoked", func(t *testing.T) {
-			cslBytes, err := json.Marshal(&cslstatus.CSL{ID: "https://example.gov/status/24", VC: []string{
-				strings.ReplaceAll(validVCStatus, "#ID", "https://issuer.oidp.uscis.gov/credentials/83627465"),
-				strings.ReplaceAll(validVCStatus, "#ID", "http://example.edu/credentials/1872")}})
-			require.NoError(t, err)
-			op.httpClient = &mockHTTPClient{doValue: &http.Response{StatusCode: http.StatusOK,
-				Body: ioutil.NopCloser(strings.NewReader(string(cslBytes)))}}
-
-			vc.Status = &verifiable.TypedID{
-				ID: "http://example.com/status/100",
-			}
-
-			vcBytes, err := vc.MarshalJSON()
-			require.NoError(t, err)
-
-			req := &CredentialsVerificationRequest{
-				Credential: vcBytes,
-				Opts: &CredentialsVerificationOptions{
-					Checks: []string{statusCheck},
-				},
-			}
-
-			reqBytes, err := json.Marshal(req)
-			require.NoError(t, err)
-
-			rr := serveHTTP(t, verificationsHandler.Handle(), http.MethodPost, endpoint, reqBytes)
-
-			require.Equal(t, http.StatusBadRequest, rr.Code)
-
-			verificationResp := &CredentialsVerificationFailResponse{}
-			err = json.Unmarshal(rr.Body.Bytes(), &verificationResp)
-			require.NoError(t, err)
-			require.Equal(t, 1, len(verificationResp.Checks))
-			require.Equal(t, statusCheck, verificationResp.Checks[0].Check)
-			require.Contains(t, verificationResp.Checks[0].Error, "Revoked")
-		})
-	})
-
-	t.Run("credential verification - invalid check", func(t *testing.T) {
-		invalidCheckName := "invalidCheckName"
-
-		req := &CredentialsVerificationRequest{
-			Credential: []byte(prCardVC),
-			Opts: &CredentialsVerificationOptions{
-				Checks: []string{invalidCheckName},
-			},
-		}
-
-		reqBytes, err := json.Marshal(req)
-		require.NoError(t, err)
-
-		rr := serveHTTP(t, verificationsHandler.Handle(), http.MethodPost, endpoint, reqBytes)
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-		verificationResp := &CredentialsVerificationFailResponse{}
-		err = json.Unmarshal(rr.Body.Bytes(), &verificationResp)
-		require.NoError(t, err)
-		require.Equal(t, 1, len(verificationResp.Checks))
-		require.Equal(t, invalidCheckName, verificationResp.Checks[0].Check)
-		require.Equal(t, "check not supported", verificationResp.Checks[0].Error)
-	})
-
-	t.Run("credential verification - invalid json input", func(t *testing.T) {
-		rr := serveHTTP(t, verificationsHandler.Handle(), http.MethodPost, endpoint,
-			[]byte("invalid input"))
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-		require.Contains(t, rr.Body.String(), "Invalid request")
-	})
-
-	t.Run("credential verification - invalid challenge and domain", func(t *testing.T) {
-		pubKey, privKey, err := ed25519.GenerateKey(rand.Reader)
-		require.NoError(t, err)
-
-		didDoc := createDIDDoc(didID, pubKey)
-		verificationMethod := didDoc.PublicKey[0].ID
-
-		op, err := New(&Config{
-			Crypto:             &cryptomock.Crypto{},
-			StoreProvider:      memstore.NewProvider(),
-			KMSSecretsProvider: mem.NewProvider(),
-			KeyManager:         &kms.KeyManager{CreateKeyValue: kh},
-			VDRI:               &vdrimock.MockVDRIRegistry{ResolveValue: didDoc},
-		})
-		require.NoError(t, err)
-
-		// verify credential
-		handler := getHandler(t, op, endpoint, verifierMode)
-
-		vReq := &CredentialsVerificationRequest{
-			Credential: getSignedVC(t, privKey, prCardVC, verificationMethod, domain,
-				"invalid-challenge"),
-			Opts: &CredentialsVerificationOptions{
-				Checks:    []string{proofCheck, statusCheck},
-				Challenge: challenge,
-				Domain:    domain,
-			},
-		}
-
-		vReqBytes, err := json.Marshal(vReq)
-		require.NoError(t, err)
-
-		rr := serveHTTP(t, handler.Handle(), http.MethodPost, endpoint, vReqBytes)
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-		require.Contains(t, rr.Body.String(), "invalid challenge in the proof")
-
-		vReq = &CredentialsVerificationRequest{
-			Credential: getSignedVC(t, privKey, prCardVC, verificationMethod, "invalid-domain", challenge),
-			Opts: &CredentialsVerificationOptions{
-				Checks:    []string{proofCheck},
-				Domain:    domain,
-				Challenge: challenge,
-			},
-		}
-
-		vReqBytes, err = json.Marshal(vReq)
-		require.NoError(t, err)
-
-		rr = serveHTTP(t, handler.Handle(), http.MethodPost, endpoint, vReqBytes)
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-		require.Contains(t, rr.Body.String(), "invalid domain in the proof")
-
-		// fail when proof has challenge and no challenge in the options
-		vReq = &CredentialsVerificationRequest{
-			Credential: getSignedVC(t, privKey, prCardVC, verificationMethod, domain, challenge),
-		}
-
-		vReqBytes, err = json.Marshal(vReq)
-		require.NoError(t, err)
-
-		rr = serveHTTP(t, handler.Handle(), http.MethodPost, endpoint, vReqBytes)
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-		require.Contains(t, rr.Body.String(), "invalid challenge in the proof")
-
-		// fail when proof has domain and no domain in the options
-		vReq = &CredentialsVerificationRequest{
-			Credential: getSignedVC(t, privKey, prCardVC, verificationMethod, domain, challenge),
-			Opts: &CredentialsVerificationOptions{
-				Checks:    []string{proofCheck},
-				Challenge: challenge,
-			},
-		}
-
-		vReqBytes, err = json.Marshal(vReq)
-		require.NoError(t, err)
-
-		rr = serveHTTP(t, handler.Handle(), http.MethodPost, endpoint, vReqBytes)
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-		require.Contains(t, rr.Body.String(), "invalid domain in the proof")
-	})
-
-	t.Run("credential verification - invalid vc proof purpose", func(t *testing.T) {
-		pubKey, privKey, err := ed25519.GenerateKey(rand.Reader)
-		require.NoError(t, err)
-
-		didDoc := createDIDDoc(didID, pubKey)
-		didDoc.AssertionMethod = nil
-		verificationMethod := didDoc.PublicKey[0].ID
-
-		ops, err := New(&Config{
-			Crypto:             &cryptomock.Crypto{},
-			StoreProvider:      memstore.NewProvider(),
-			KMSSecretsProvider: mem.NewProvider(),
-			KeyManager:         &kms.KeyManager{CreateKeyValue: kh},
-			VDRI:               &vdrimock.MockVDRIRegistry{ResolveValue: didDoc},
-		})
-		require.NoError(t, err)
-
-		cslBytes, err := json.Marshal(&cslstatus.CSL{})
-		require.NoError(t, err)
-
-		ops.httpClient = &mockHTTPClient{doValue: &http.Response{StatusCode: http.StatusOK,
-			Body: ioutil.NopCloser(strings.NewReader(string(cslBytes)))}}
-
-		vc.Status = &verifiable.TypedID{
-			ID:   "http://example.com/status/100",
-			Type: "CredentialStatusList2017",
-		}
-
-		vcBytes, err := vc.MarshalJSON()
-		require.NoError(t, err)
-
-		// verify credential
-		handler := getHandler(t, ops, endpoint, verifierMode)
-
-		vReq := &CredentialsVerificationRequest{
-			Credential: getSignedVC(t, privKey, string(vcBytes), verificationMethod, domain, challenge),
-			Opts: &CredentialsVerificationOptions{
-				Checks:    []string{proofCheck, statusCheck},
-				Challenge: challenge,
-				Domain:    domain,
-			},
-		}
-
-		vReqBytes, err := json.Marshal(vReq)
-		require.NoError(t, err)
-
-		rr := serveHTTP(t, handler.Handle(), http.MethodPost, endpoint, vReqBytes)
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-		require.Contains(t, rr.Body.String(), "verifiable credential proof purpose validation error :"+
-			" unable to find matching assertionMethod key IDs for given verification method")
-	})
-}
-
-func TestVerifyPresentation(t *testing.T) {
-	kh, err := keyset.NewHandle(ecdhes.ECDHES256KWAES256GCMKeyTemplate())
-	require.NoError(t, err)
-
-	op, err := New(&Config{
-		Crypto:             &cryptomock.Crypto{},
-		StoreProvider:      memstore.NewProvider(),
-		KMSSecretsProvider: mem.NewProvider(),
-		KeyManager:         &kms.KeyManager{CreateKeyValue: kh},
-		VDRI:               &vdrimock.MockVDRIRegistry{},
-	})
-	require.NoError(t, err)
-
-	endpoint := presentationsVerificationEndpoint
-	verificationsHandler := getHandler(t, op, endpoint, verifierMode)
-
-	t.Run("presentation verification - success", func(t *testing.T) {
-		pubKey, privKey, err := ed25519.GenerateKey(rand.Reader)
-		require.NoError(t, err)
-
-		didID := "did:test:EiBNfNRaz1Ll8BjVsbNv-fWc7K_KIoPuW8GFCh1_Tz_Iuw=="
-
-		didDoc := createDIDDoc(didID, pubKey)
-		verificationMethod := didDoc.PublicKey[0].ID
-
-		kh, err := keyset.NewHandle(ecdhes.ECDHES256KWAES256GCMKeyTemplate())
-		require.NoError(t, err)
-
-		op, err := New(&Config{
-			Crypto:             &cryptomock.Crypto{},
-			StoreProvider:      memstore.NewProvider(),
-			KMSSecretsProvider: mem.NewProvider(),
-			KeyManager:         &kms.KeyManager{CreateKeyValue: kh},
-			VDRI:               &vdrimock.MockVDRIRegistry{ResolveValue: didDoc},
-		})
-		require.NoError(t, err)
-
-		// verify credential
-		handler := getHandler(t, op, endpoint, verifierMode)
-
-		vReq := &VerifyPresentationRequest{
-			Presentation: getSignedVP(t, privKey, prCardVC, verificationMethod, domain, challenge),
-			Opts: &VerifyPresentationOptions{
-				Checks:    []string{proofCheck},
-				Challenge: challenge,
-				Domain:    domain,
-			},
-		}
-
-		vReqBytes, err := json.Marshal(vReq)
-		require.NoError(t, err)
-
-		rr := serveHTTP(t, handler.Handle(), http.MethodPost, endpoint, vReqBytes)
-
-		require.Equal(t, http.StatusOK, rr.Code)
-
-		verificationResp := &VerifyPresentationSuccessResponse{}
-		err = json.Unmarshal(rr.Body.Bytes(), &verificationResp)
-		require.NoError(t, err)
-		require.Equal(t, 1, len(verificationResp.Checks))
-		require.Equal(t, proofCheck, verificationResp.Checks[0])
-	})
-
-	t.Run("presentation verification - request doesn't contain checks", func(t *testing.T) {
-		req := &VerifyPresentationRequest{
-			Presentation: []byte(vpWithoutProof),
-		}
-
-		reqBytes, err := json.Marshal(req)
-		require.NoError(t, err)
-
-		rr := serveHTTP(t, verificationsHandler.Handle(), http.MethodPost, endpoint, reqBytes)
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-
-		// verify that the default check was performed
-		verificationResp := &VerifyPresentationFailureResponse{}
-		err = json.Unmarshal(rr.Body.Bytes(), &verificationResp)
-		require.NoError(t, err)
-		require.Equal(t, 1, len(verificationResp.Checks))
-		require.Equal(t, proofCheck, verificationResp.Checks[0].Check)
-		require.Equal(t, "verifiable presentation proof validation error : embedded proof is missing",
-			verificationResp.Checks[0].Error)
-	})
-
-	t.Run("presentation verification - proof check failure", func(t *testing.T) {
-		// no proof in VC
-		req := &VerifyPresentationRequest{
-			Presentation: []byte(vpWithoutProof),
-			Opts: &VerifyPresentationOptions{
-				Checks: []string{proofCheck},
-			},
-		}
-
-		reqBytes, err := json.Marshal(req)
-		require.NoError(t, err)
-
-		rr := serveHTTP(t, verificationsHandler.Handle(), http.MethodPost, endpoint, reqBytes)
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-
-		verificationResp := &VerifyPresentationFailureResponse{}
-		err = json.Unmarshal(rr.Body.Bytes(), &verificationResp)
-		require.NoError(t, err)
-		require.Equal(t, 1, len(verificationResp.Checks))
-		require.Equal(t, proofCheck, verificationResp.Checks[0].Check)
-		require.Equal(t, "verifiable presentation proof validation error : embedded proof is missing",
-			verificationResp.Checks[0].Error)
-
-		// proof validation error (DID not found)
-		req = &VerifyPresentationRequest{
-			Presentation: []byte(validVCWithProof),
-			Opts: &VerifyPresentationOptions{
-				Checks: []string{proofCheck},
-			},
-		}
-
-		reqBytes, err = json.Marshal(req)
-		require.NoError(t, err)
-
-		rr = serveHTTP(t, verificationsHandler.Handle(), http.MethodPost, endpoint, reqBytes)
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-
-		verificationResp = &VerifyPresentationFailureResponse{}
-		err = json.Unmarshal(rr.Body.Bytes(), &verificationResp)
-		require.NoError(t, err)
-		require.Equal(t, 1, len(verificationResp.Checks))
-		require.Equal(t, proofCheck, verificationResp.Checks[0].Check)
-		require.Contains(t, verificationResp.Checks[0].Error, "proof validation error")
-	})
-
-	t.Run("presentation verification - invalid check", func(t *testing.T) {
-		invalidCheckName := "invalidCheckName"
-
-		req := &VerifyPresentationRequest{
-			Presentation: []byte(validVP),
-			Opts: &VerifyPresentationOptions{
-				Checks: []string{invalidCheckName},
-			},
-		}
-
-		reqBytes, err := json.Marshal(req)
-		require.NoError(t, err)
-
-		rr := serveHTTP(t, verificationsHandler.Handle(), http.MethodPost, endpoint, reqBytes)
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-		verificationResp := &VerifyPresentationFailureResponse{}
-		err = json.Unmarshal(rr.Body.Bytes(), &verificationResp)
-		require.NoError(t, err)
-		require.Equal(t, 1, len(verificationResp.Checks))
-		require.Equal(t, invalidCheckName, verificationResp.Checks[0].Check)
-		require.Equal(t, "check not supported", verificationResp.Checks[0].Error)
-	})
-
-	t.Run("presentation verification - invalid json input", func(t *testing.T) {
-		rr := serveHTTP(t, verificationsHandler.Handle(), http.MethodPost, endpoint,
-			[]byte("invalid input"))
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-		require.Contains(t, rr.Body.String(), "Invalid request")
-	})
-
-	t.Run("presentation verification - invalid challenge and domain", func(t *testing.T) {
-		pubKey, privKey, err := ed25519.GenerateKey(rand.Reader)
-		require.NoError(t, err)
-
-		didID := "did:test:xyz"
-
-		didDoc := createDIDDoc(didID, pubKey)
-		verificationMethod := didDoc.PublicKey[0].ID
-
-		kh, err := keyset.NewHandle(ecdhes.ECDHES256KWAES256GCMKeyTemplate())
-		require.NoError(t, err)
-
-		op, err := New(&Config{
-			Crypto:             &cryptomock.Crypto{},
-			StoreProvider:      memstore.NewProvider(),
-			KMSSecretsProvider: mem.NewProvider(),
-			KeyManager:         &kms.KeyManager{CreateKeyValue: kh},
-			VDRI:               &vdrimock.MockVDRIRegistry{ResolveValue: didDoc},
-		})
-		require.NoError(t, err)
-
-		// verify credential
-		handler := getHandler(t, op, endpoint, verifierMode)
-
-		vReq := &VerifyPresentationRequest{
-			Presentation: getSignedVP(t, privKey, prCardVC, verificationMethod, domain, uuid.New().String()),
-			Opts: &VerifyPresentationOptions{
-				Checks:    []string{proofCheck},
-				Domain:    domain,
-				Challenge: challenge,
-			},
-		}
-
-		vReqBytes, err := json.Marshal(vReq)
-		require.NoError(t, err)
-
-		rr := serveHTTP(t, handler.Handle(), http.MethodPost, endpoint, vReqBytes)
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-		require.Contains(t, rr.Body.String(), "invalid challenge in the proof")
-
-		vReq = &VerifyPresentationRequest{
-			Presentation: getSignedVP(t, privKey, prCardVC, verificationMethod, "invalid-domain", challenge),
-			Opts: &VerifyPresentationOptions{
-				Checks:    []string{proofCheck},
-				Domain:    domain,
-				Challenge: challenge,
-			},
-		}
-
-		vReqBytes, err = json.Marshal(vReq)
-		require.NoError(t, err)
-
-		rr = serveHTTP(t, handler.Handle(), http.MethodPost, endpoint, vReqBytes)
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-		require.Contains(t, rr.Body.String(), "invalid domain in the proof")
-
-		// fail when proof has challenge and no challenge in the options
-		vReq = &VerifyPresentationRequest{
-			Presentation: getSignedVP(t, privKey, prCardVC, verificationMethod, domain, challenge),
-		}
-
-		vReqBytes, err = json.Marshal(vReq)
-		require.NoError(t, err)
-
-		rr = serveHTTP(t, handler.Handle(), http.MethodPost, endpoint, vReqBytes)
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-		require.Contains(t, rr.Body.String(), "invalid challenge in the proof")
-
-		// fail when proof has domain and no domain in the options
-		vReq = &VerifyPresentationRequest{
-			Presentation: getSignedVP(t, privKey, prCardVC, verificationMethod, domain, challenge),
-			Opts: &VerifyPresentationOptions{
-				Checks:    []string{proofCheck},
-				Challenge: challenge,
-			},
-		}
-
-		vReqBytes, err = json.Marshal(vReq)
-		require.NoError(t, err)
-
-		rr = serveHTTP(t, handler.Handle(), http.MethodPost, endpoint, vReqBytes)
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-		require.Contains(t, rr.Body.String(), "invalid domain in the proof")
-	})
-
-	t.Run("presentation verification - invalid vp proof purpose", func(t *testing.T) {
-		pubKey, privKey, err := ed25519.GenerateKey(rand.Reader)
-		require.NoError(t, err)
-
-		didID := "did:test:xyz123"
-
-		didDoc := createDIDDoc(didID, pubKey)
-		didDoc.Authentication = nil
-		verificationMethod := didDoc.PublicKey[0].ID
-
-		kh, err := keyset.NewHandle(ecdhes.ECDHES256KWAES256GCMKeyTemplate())
-		require.NoError(t, err)
-
-		op, err := New(&Config{
-			Crypto:             &cryptomock.Crypto{},
-			StoreProvider:      memstore.NewProvider(),
-			KMSSecretsProvider: mem.NewProvider(),
-			KeyManager:         &kms.KeyManager{CreateKeyValue: kh},
-			VDRI:               &vdrimock.MockVDRIRegistry{ResolveValue: didDoc},
-		})
-		require.NoError(t, err)
-
-		// verify credential
-		handler := getHandler(t, op, endpoint, verifierMode)
-
-		vReq := &VerifyPresentationRequest{
-			Presentation: getSignedVP(t, privKey, prCardVC, verificationMethod, domain, challenge),
-			Opts: &VerifyPresentationOptions{
-				Checks:    []string{proofCheck},
-				Challenge: challenge,
-				Domain:    domain,
-			},
-		}
-
-		vReqBytes, err := json.Marshal(vReq)
-		require.NoError(t, err)
-
-		rr := serveHTTP(t, handler.Handle(), http.MethodPost, endpoint, vReqBytes)
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-		require.Contains(t, rr.Body.String(), "verifiable presentation proof purpose validation error :"+
-			" unable to find matching authentication key IDs for given verification method")
-	})
-
-	t.Run("presentation verification - invalid vc proof purpose", func(t *testing.T) {
-		pubKey, privKey, err := ed25519.GenerateKey(rand.Reader)
-		require.NoError(t, err)
-
-		didID := "did:test:abc123"
-
-		didDoc := createDIDDoc(didID, pubKey)
-		didDoc.AssertionMethod = nil
-		verificationMethod := didDoc.PublicKey[0].ID
-
-		kh, err := keyset.NewHandle(ecdhes.ECDHES256KWAES256GCMKeyTemplate())
-		require.NoError(t, err)
-
-		op, err := New(&Config{
-			Crypto:             &cryptomock.Crypto{},
-			StoreProvider:      memstore.NewProvider(),
-			KMSSecretsProvider: mem.NewProvider(),
-			KeyManager:         &kms.KeyManager{CreateKeyValue: kh},
-			VDRI:               &vdrimock.MockVDRIRegistry{ResolveValue: didDoc},
-		})
-		require.NoError(t, err)
-
-		// verify credential
-		handler := getHandler(t, op, endpoint, verifierMode)
-
-		vReq := &VerifyPresentationRequest{
-			Presentation: getSignedVP(t, privKey, prCardVC, verificationMethod, domain, challenge),
-			Opts: &VerifyPresentationOptions{
-				Checks:    []string{proofCheck},
-				Challenge: challenge,
-				Domain:    domain,
-			},
-		}
-
-		vReqBytes, err := json.Marshal(vReq)
-		require.NoError(t, err)
-
-		rr := serveHTTP(t, handler.Handle(), http.MethodPost, endpoint, vReqBytes)
-
-		require.Equal(t, http.StatusBadRequest, rr.Code)
-		require.Contains(t, rr.Body.String(), "verifiable credential proof purpose validation error : unable"+
-			" to find matching assertionMethod key IDs for given verification method")
-	})
-}
-
-func TestValidateProof(t *testing.T) {
-	proof := make(map[string]interface{})
-	key := "challenge"
-	value := uuid.New().String()
-
-	proof[key] = value
-
-	// success
-	err := validateProofData(proof, key, value)
-	require.NoError(t, err)
-
-	// fail - not a string
-	proof[key] = 234
-	err = validateProofData(proof, key, value)
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "invalid challenge in the proof")
-
-	// fail - invalid
-	proof[key] = "invalid-data"
-	err = validateProofData(proof, key, value)
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "invalid challenge in the proof")
-}
-
-func TestValidateProofPurpose(t *testing.T) {
-	pubKey, _, err := ed25519.GenerateKey(rand.Reader)
-	require.NoError(t, err)
-
-	didID := "did:test:xyz123"
-
-	didDoc := createDIDDoc(didID, pubKey)
-	kid := didDoc.PublicKey[0].ID
-
-	vdriReg := &vdrimock.MockVDRIRegistry{ResolveValue: didDoc}
-
-	proof := make(map[string]interface{})
-	key := "challenge"
-	value := uuid.New().String()
-
-	proof[proofPurpose] = assertionMethod
-	proof[verificationMethod] = kid
-
-	// success
-	err = validateProofPurpose(proof, vdriReg)
-	require.NoError(t, err)
-
-	// fail - no value
-	delete(proof, proofPurpose)
-	err = validateProofPurpose(proof, vdriReg)
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "proof doesn't have purpose")
-
-	proof[proofPurpose] = assertionMethod
-	delete(proof, verificationMethod)
-	err = validateProofPurpose(proof, vdriReg)
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "proof doesn't have verification method")
-
-	// fail - not a string
-	proof[proofPurpose] = 234
-	err = validateProofPurpose(proof, vdriReg)
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "proof purpose is not a string")
-
-	proof[proofPurpose] = assertionMethod
-	proof[verificationMethod] = 234
-	err = validateProofPurpose(proof, vdriReg)
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "proof verification method is not a string")
-
-	// fail - invalid
-	proof[key] = "invalid-data"
-	err = validateProofData(proof, key, value)
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "invalid challenge in the proof")
 }
 
 func serveHTTP(t *testing.T, handler http.HandlerFunc, method, path string, req []byte) *httptest.ResponseRecorder {
@@ -3305,16 +2303,16 @@ func getProfileRequest() *ProfileRequest {
 		SignatureType: "Ed25519Signature2018"}
 }
 
-func getHandler(t *testing.T, op *Operation, lookup, mode string) Handler {
-	return getHandlerWithError(t, op, lookup, mode)
+func getHandler(t *testing.T, op *Operation, lookup string) Handler {
+	return getHandlerWithError(t, op, lookup)
 }
 
-func getHandlerWithError(t *testing.T, op *Operation, lookup, mode string) Handler {
-	return handlerLookup(t, op, lookup, mode)
+func getHandlerWithError(t *testing.T, op *Operation, lookup string) Handler {
+	return handlerLookup(t, op, lookup)
 }
 
-func handlerLookup(t *testing.T, op *Operation, lookup, mode string) Handler {
-	handlers, err := op.GetRESTHandlers(mode)
+func handlerLookup(t *testing.T, op *Operation, lookup string) Handler {
+	handlers, err := op.GetRESTHandlers()
 	require.NoError(t, err)
 	require.NotEmpty(t, handlers)
 
@@ -3500,93 +2498,6 @@ func (m *mockVCStatusManager) UpdateVCStatus(v *verifiable.Credential, profile *
 
 func (m *mockVCStatusManager) GetCSL(id string) (*cslstatus.CSL, error) {
 	return m.getCSLValue, m.getCSLErr
-}
-
-type mockHTTPClient struct {
-	doValue *http.Response
-	doErr   error
-}
-
-func (m *mockHTTPClient) Do(req *http.Request) (*http.Response, error) {
-	return m.doValue, m.doErr
-}
-
-func getSignedVC(t *testing.T, privKey []byte, vcJSON, verificationMethod, domain, challenge string) []byte {
-	vc, err := verifiable.NewUnverifiedCredential([]byte(vcJSON))
-	require.NoError(t, err)
-
-	created, err := time.Parse(time.RFC3339, "2018-03-15T00:00:00Z")
-	require.NoError(t, err)
-
-	signerSuite := ed25519signature2018.New(
-		suite.WithSigner(getEd25519TestSigner(privKey)),
-		suite.WithCompactProof())
-	err = vc.AddLinkedDataProof(&verifiable.LinkedDataProofContext{
-		SignatureType:           "Ed25519Signature2018",
-		Suite:                   signerSuite,
-		SignatureRepresentation: verifiable.SignatureJWS,
-		Created:                 &created,
-		VerificationMethod:      verificationMethod,
-		Domain:                  domain,
-		Challenge:               challenge,
-		Purpose:                 vccrypto.AssertionMethod,
-	})
-	require.NoError(t, err)
-
-	require.Len(t, vc.Proofs, 1)
-
-	signedVC, err := vc.MarshalJSON()
-	require.NoError(t, err)
-
-	return signedVC
-}
-
-func getSignedVP(t *testing.T, privKey []byte, vcJSON, verificationMethod, domain, challenge string) []byte { // nolint
-	signedVC := getSignedVC(t, privKey, vcJSON, verificationMethod, "", "")
-
-	vc, err := verifiable.NewUnverifiedCredential(signedVC)
-	require.NoError(t, err)
-
-	created, err := time.Parse(time.RFC3339, "2018-03-15T00:00:00Z")
-	require.NoError(t, err)
-
-	vp, err := vc.Presentation()
-	require.NoError(t, err)
-
-	signerSuite := ed25519signature2018.New(
-		suite.WithSigner(getEd25519TestSigner(privKey)),
-		suite.WithCompactProof())
-	err = vp.AddLinkedDataProof(&verifiable.LinkedDataProofContext{
-		SignatureType:           "Ed25519Signature2018",
-		Suite:                   signerSuite,
-		SignatureRepresentation: verifiable.SignatureJWS,
-		Created:                 &created,
-		VerificationMethod:      verificationMethod,
-		Domain:                  domain,
-		Challenge:               challenge,
-		Purpose:                 vccrypto.Authentication,
-	})
-	require.NoError(t, err)
-
-	signedVP, err := vp.MarshalJSON()
-	require.NoError(t, err)
-
-	return signedVP
-}
-
-type ed25519TestSigner struct {
-	privateKey []byte
-}
-
-func (s *ed25519TestSigner) Sign(doc []byte) ([]byte, error) {
-	if l := len(s.privateKey); l != ed25519.PrivateKeySize {
-		return nil, errors.New("ed25519: bad private key length")
-	}
-
-	return ed25519.Sign(s.privateKey, doc), nil
-}
-func getEd25519TestSigner(privKey []byte) *ed25519TestSigner {
-	return &ed25519TestSigner{privateKey: privKey}
 }
 
 type mockCredentialStatusManager struct {
